EmuHawk: PAVirtualListView - finish consuming InputRoll code, tidy and remove un-needed things (lagframes, horizontal rotation etc)
