+++ BizHawk.Client.EmuHawk/BizHawk.Client.EmuHawk.csproj
@@ -601,7 +601,28 @@
     <Compile Include="CustomControls\MsgBox.designer.cs">
       <DependentUpon>MsgBox.cs</DependentUpon>
     </Compile>
+    <Compile Include="CustomControls\PlatformAgnosticVirtualListView.API.cs">
+      <DependentUpon>PlatformAgnosticVirtualListView.cs</DependentUpon>
+      <SubType>Component</SubType>
+    </Compile>
+    <Compile Include="CustomControls\PlatformAgnosticVirtualListView.Classes.cs">
+      <DependentUpon>PlatformAgnosticVirtualListView.cs</DependentUpon>
+      <SubType>Component</SubType>
+    </Compile>
     <Compile Include="CustomControls\PlatformAgnosticVirtualListView.Drawing.cs">
+      <DependentUpon>PlatformAgnosticVirtualListView.cs</DependentUpon>
+      <SubType>Component</SubType>
+    </Compile>
+    <Compile Include="CustomControls\PlatformAgnosticVirtualListView.EventHandlers.cs">
+      <DependentUpon>PlatformAgnosticVirtualListView.cs</DependentUpon>
+      <SubType>Component</SubType>
+    </Compile>
+    <Compile Include="CustomControls\PlatformAgnosticVirtualListView.Helpers.cs">
+      <DependentUpon>PlatformAgnosticVirtualListView.cs</DependentUpon>
+      <SubType>Component</SubType>
+    </Compile>
+    <Compile Include="CustomControls\PlatformAgnosticVirtualListView.Properties.cs">
+      <DependentUpon>PlatformAgnosticVirtualListView.cs</DependentUpon>
       <SubType>Component</SubType>
     </Compile>
     <Compile Include="CustomControls\PrereqsAlert.cs">
